requirements.txt
streamlit==0.74.1
h5py==2.10.0
matplotlib==3.3.3
opencv_contrib_python==4.5.1.48
numpy==1.19.3
dlib==19.21.1
tensorflow==2.4.0
pandas==1.2.0
Keras==2.4.3
imutils==0.5.3
Pillow==8.1.0
